requirements.txt
tensorflow>=2.12.0
numpy>=1.23.0
matplotlib>=3.6.0
Pillow>=9.4.0
scikit-learn>=1.2.0
tqdm>=4.65.0
streamlit>=1.22.0
pandas>=1.5.0
requests>=2.28.0
kaggle>=1.5.13
